Payment

+++ store/models.py
@@ -1,3 +1,4 @@
+from unicodedata import decimal
 from django.db import models
 from django.contrib.auth.models import User
 
@@ -14,7 +15,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
 	name = models.CharField(max_length=200)
-	price = models.FloatField()
+	price = models.DecimalField(max_digits=6, decimal_places= 2 )
 	image = models.ImageField(null=True, blank=True)
 
 	def __str__(self):
@@ -44,7 +45,6 @@ class Order(models.Model):
 		for i in orderitems:
 				shipping = True
 		return shipping
-
 	@property
 	def get_cart_total(self):
 		orderitems = self.orderitem_set.all()
